Define some terminology
Terms like should et. al. are commonly understood to introduce a
requirement, but it is sometimes unclear exactly what they mean. Link to
the RFC specifying particular meanings for those terms.

The term "hart" is used extensively without definition. The idea of an
EEI is used in the following commit TODO.

What exactly constitutes a valid **hartid** was not previously
explicitly defined.

Signed-off-by: Sean Anderson <seanga2@gmail.com>