Add Hart Discovery Extension
This extension is an attempt to make it easier for supervisor mode
software to identify which harts are valid, and which are part of the
execution environment.

There is currently no way to consistently discover which harts are
valid, except by calling a function like sbi_send_ipi for ALL possible
values of hart_mask, and then checking to see if SBI_ERR_INVALID_PARAM
is returned. If it is, then repeated calls can bisect which harts are
valid. However, this is slow and inefficient.

One alternate method could be to allocate a bit vector, and have each
hart set a bit when it boots. However, the size of the vector cannot be
determined in advance. This risks data corruption if (e.g.) the vector
is 64 bits long and hart 65 tries to set its bit. With the function
sbi_get_max_hartid, a suitable vector may be allocated in advance. If
dynamic allocation is not possible/desired, it can also act as a
sanity-check to ensure the pre-allocated size is not exceeded.

With the HSM extension, a supervisor may wish to determine which harts
can be stopped and started. Without sbi_get_max_hartid, there is no way
to identify harts which are stopped on boot but which can be started
(except by trying ALL possible values of hartid). Even with this
function, determining which hartids are valid requires callind
sbi_hart_status for every hart from 0 to the maximum hartid. In an
execution environment with a large maximum hartid but few valid harts,
this can be inefficient. sbi_get_valid_harts allows supervisors to only
check the status of harts which are valid.

Signed-off-by: Sean Anderson <seanga2@gmail.com>